+++ edison-status/src/main/java/de/otto/edison/status/domain/ApplicationStatus.java
@@ -10,46 +10,46 @@ import static java.util.Collections.unmodifiableList;
 @Immutable
 public final class ApplicationStatus {
 
-    private final String name;
+    private final ApplicationInfo applicationInfo;
+    private final SystemInfo systemInfo;
+    private final VersionInfo versionInfo;
     private final Status status;
     private final List<StatusDetail> statusDetails;
-    private final VersionInfo versionInfo;
-    private final String hostName;
 
-    private ApplicationStatus(final String applicationName,
-                              final String hostName,
+    private ApplicationStatus(final ApplicationInfo applicationInfo,
+                              final SystemInfo systemInfo,
                               final VersionInfo versionInfo,
                               final List<StatusDetail> details) {
-        this.name = applicationName;
-        this.versionInfo = versionInfo;
-        this.hostName = hostName;
         this.status = details.stream()
                 .map(StatusDetail::getStatus)
                 .reduce(Status.OK, Status::plus);
         this.statusDetails = unmodifiableList(new ArrayList<>(details));
+        this.applicationInfo = applicationInfo;
+        this.systemInfo = systemInfo;
+        this.versionInfo = versionInfo;
     }
 
-    public static ApplicationStatus applicationStatus(final String applicationName,
-                                                      final String hostName,
+    public static ApplicationStatus applicationStatus(final ApplicationInfo applicationInfo,
+                                                      final SystemInfo systemInfo,
                                                       final VersionInfo versionInfo,
                                                       final List<StatusDetail> details) {
-        return new ApplicationStatus(applicationName, hostName, versionInfo, details);
+        return new ApplicationStatus(applicationInfo, systemInfo, versionInfo, details);
     }
 
-    public String getName() {
-        return name;
+    public ApplicationInfo getApplicationInfo() {
+        return applicationInfo;
     }
 
-    public Status getStatus() {
-        return status;
-    }
-
-    public String getHostName() {
-        return hostName;
+    public SystemInfo getSystemInfo() {
+        return systemInfo;
     }
 
     public VersionInfo getVersionInfo() {
         return versionInfo;
+    }
+
+    public Status getStatus() {
+        return status;
     }
 
     public List<StatusDetail> getStatusDetails() {
@@ -63,34 +63,33 @@ public final class ApplicationStatus {
 
         ApplicationStatus that = (ApplicationStatus) o;
 
-        if (hostName != null ? !hostName.equals(that.hostName) : that.hostName != null) return false;
-        if (name != null ? !name.equals(that.name) : that.name != null) return false;
-        if (status != that.status) return false;
-        if (statusDetails != null ? !statusDetails.equals(that.statusDetails) : that.statusDetails != null)
+        if (applicationInfo != null ? !applicationInfo.equals(that.applicationInfo) : that.applicationInfo != null)
             return false;
+        if (systemInfo != null ? !systemInfo.equals(that.systemInfo) : that.systemInfo != null) return false;
         if (versionInfo != null ? !versionInfo.equals(that.versionInfo) : that.versionInfo != null) return false;
+        if (status != that.status) return false;
+        return !(statusDetails != null ? !statusDetails.equals(that.statusDetails) : that.statusDetails != null);
 
-        return true;
     }
 
     @Override
     public int hashCode() {
-        int result = name != null ? name.hashCode() : 0;
+        int result = applicationInfo != null ? applicationInfo.hashCode() : 0;
+        result = 31 * result + (systemInfo != null ? systemInfo.hashCode() : 0);
+        result = 31 * result + (versionInfo != null ? versionInfo.hashCode() : 0);
         result = 31 * result + (status != null ? status.hashCode() : 0);
         result = 31 * result + (statusDetails != null ? statusDetails.hashCode() : 0);
-        result = 31 * result + (versionInfo != null ? versionInfo.hashCode() : 0);
-        result = 31 * result + (hostName != null ? hostName.hashCode() : 0);
         return result;
     }
 
     @Override
     public String toString() {
         return "ApplicationStatus{" +
-                "hostName='" + hostName + '\'' +
-                ", name='" + name + '\'' +
+                "applicationInfo=" + applicationInfo +
+                ", systemInfo=" + systemInfo +
+                ", versionInfo=" + versionInfo +
                 ", status=" + status +
                 ", statusDetails=" + statusDetails +
-                ", versionInfo=" + versionInfo +
                 '}';
     }
 }